[RFC] Remove multi-level wishbone muxing
[We need to make sure this is worthwhile ... seems to improve
 timing a bit but needs more testing ]

The LiteDRAM wrapper would have its own muxes on the wishbone
to split between DRAM, CSR and BIOS. This makes it harder to
control overall timing and separate fast and slow path.

Instead do it all inside a single mux in the SOC interconnect

This seems to also very slightly improve timing already.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>