Generate data storage interrupts
This adds a path from loadstore1 back to execute1 for reporting
errors and the machinery in execute1 for writing DAR and DSISR
and generating data storage interrupts at vector 0x300.

If dcache is given two requests in successive cycles and the
first encounters an error (e.g. a TLB miss), it will now cancel
the second request.

Loadstore1 now responds to errors reported by dcache by sending
the exception info (address and error bits) to execute1 and
returning to the idle state.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>