Pass mtspr/mfspr to MMU-related SPRs down to loadstore1
This arranges for some mfspr and mtspr to get sent to loadstore1
instead of being handled in execute1.  In particular, DAR and DSISR
are handled this way.  They therefore become "slow" SPRs rather than
fast ones.  This then means that the writing of DAR and DSISR on a
DSI happens entirely within loadstore1 instead of those values
having to be sent over to execute1, and in delivering the DSI,
execute1 only has to write SRR0/1 (as for other interrupts) and
doesn't need the logic to write back DAR and DSISR to the register
file.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>