MMU: Add a register in the shifter to help make timing
At present we have one cycle to get from the shift count in r.shift
through the address shifter to the address sent to the dcache, then
through a multiplexer into the cache tag and TLB RAMs.  This seems
to be a little tight on the A7-100 at 100MHz.  To alleviate this,
we move the first stage of the shifter into the previous cycle by
introducing a register between the first and second multiplexers
in the shifter.  With this we make timing on the A7-100 at 100MHz.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>