Add TLB to icache
This adds a direct-mapped TLB to the icache, with 64 entries by default.
Execute1 now sends a "virt_mode" signal from MSR[IR] to fetch1 along
with redirects to indicate whether instruction addresses should be
translated through the TLB, and fetch1 sends that on to icache.
Similarly a "priv_mode" signal is sent to indicate the privilege
mode for instruction fetches.  This means that changes to MSR[IR]
or MSR[PR] don't take effect until the next redirect, meaning an
isync, rfid, branch, etc.

The icache uses a hash of the effective address (i.e. next instruction
address) to index the TLB.  The hash is an XOR of three fields of the
address; with a 64-entry TLB, the fields are bits 12--17, 18--23 and
24--29 of the address.  TLB invalidations simply invalidate the
indexed TLB entry without checking the contents.

If the icache detects a TLB miss with virt_mode=1, it will send a
single fetch_failed indication through fetch2 to decode1, which will
turn it into a special OP_FETCH_FAILED opcode with unit=LDST.  That
will get sent down to loadstore1 which will currently just raise
a Instruction Storage Interrupt (0x400) exception.

Three bits in the PTE obtained from the TLB are used to check whether
an instruction access is allowed -- the execute permission bit (bit
0), the privilege bit (bit 3) and the reference bit (bit 8).  If bit 0
is 0, bit 8 is 0, or bit 3 is 1 and priv_mode=0, then a fetch_failed
indication is sent down to fetch2 and to decode1, which generates an
OP_FETCH_FAILED with unit=ALU.  That goes down to execute1, which
generates an ISI with SRR1 bits set as appropriate.

Tlbie operations get sent from mmu to icache over a new connection.
For testing purposes, a tlbie with RB(11:10) = 00 and RB(8) = 1 will
create a valid TLB entry from the contents of RS.  Mmu can also send
an explicit TLB load operation, but that is not used yet.

Now that we have instruction translation, the privileged instruction
test needs to create TLB entries, since it executes code in problem
state.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>