Filter out duplicate fetch failure in decode1 not icache
We had a hang where a fetch failure signal got lost in fetch2 due
to a stall ending, meaning that the icache output got overwritten
with the stash buffer.  It happened that icache was signalling a
fetch failure in that cycle, and since it only signals it once, it
got lost.

Instead, icache now signals the fetch failure continuously until
it gets a TLB reload or flush, and decode1 filters that down to
a single OP_FETCH_FAILED.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>