dcache: Fix TLB lookup for second pipelined access
This fixes a bug that happens when two accesses are submitted in
successive cycles and the first causes a stall.  When the stall
ends and the second request gets processed, we were using the
wrong TLB data because the TLB was read using d_in.addr from the
previous cycle, not the address originally presented.  This fixes
it by keeping the address constant while the stall is present.

Fixes: 278c3076d2db ("dcache: Implement data TLB", 2020-04-06)
Signed-off-by: Paul Mackerras <paulus@ozlabs.org>