core: Start of VMX/VSX support
This extends the register file to 192 entries in order to hold vector
registers plus the extra 32 doublewords required to complete the VSX
register set.  The gspr_index_t is extended to 8 bits, and the FPRs
are now entries 64, 66, ... 126 to allow the extra registers added
with VSX to sit next to the corresponding FPRs.

This adds new values VEC and VSX for the facility field of the decode
tables.  In fact this won't be quite sufficient as there are some VSX
instructions which check either MSR_VEC or MSR_VSX depending on which
VSR is being used.

This does not add any new instructions yet.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>