core: Implement VSX load-with-length and store-with-length instructions
This implements the lxvl, lxvll, stxvl and stxvll instructions.

This means that loadstore1 needs to be able to cope with any length
from 0 to 8 bytes for each transfer.  For lxvll and stxvll, and for
lxvl and stxvl in big-endian mode, the byte-swapping logic needs to
left-justify the result in the register, meaning that the byte-swap is
always done like an 8-byte byte-swap regardless of the transfer
length.  Furthermore the trimming logic for reads needs to enable
bytes on the left rather than the right.

This adds tests for lxvl, lxvll, stxvl and stxvll at 8 different
lengths.  (Testing all 17 values from 0 to 16 made the vsx test
take over 40 seconds in simulation, so the set of values was
reduced to 8.)

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>