core: Implement vperm and vpermr instructions
And extend the vector test to test them too.

These instructions are done in two halves, left then right.  However,
the result for the left half of the vector register may need bytes
from the right half of the sources.  Therefore the first half
instruction stores its operands and raises the busy signal until the
second half of the instruction appears in the execute1 input
register.  The first half can then use the right halves of VRA and VRB
as well as the left halves which were saved away.  The right half
instruction can then also access both halves of the VRA and VRB
operands.  The whole thing takes 3 cycles.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>