decode1: Decode individual vector instructions
Previously, vector instructions were decoded in groups of 32, which is
fine for many things but does not give enough resolution for things
such as mtvscr and mfvscr.  It turns out to be sufficient to decode
individual instructions for columns 0 - 15 and to decode by column for
columns 16 - 63 (where "columns" are indexed by bits 5..0 of the
instruction word).

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>