core: Implement xxpermdi instruction
This uses the permutation machinery.

This is the first instruction to be implemented with major opcode 60.
This major opcode has a 10-bit minor opcode, but most instructions are
in groups of 4, hence the decode table uses bits 10..3 (organized as
5..3 followed by 10..6).  For those cases where a group of 4
instructions are not all valid and not all invalid, there is a 1-bit
table indexed by all 10 bits to indicate which instructions are
invalid.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>