loadstore: Convert to 3-stage pipeline
This makes loadstore use a 3-stage pipeline.  For now, only one
instruction goes through the pipe at a time, and completion and
writeback are now one cycle later than before, so performance is down
a bit.  In future it should be able to sustain one load or store per
cycle provided they hit in the dcache.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>