core: Track GPR hazards using tags that propagate through the pipelines
This changes the way GPR hazards are detected and tracked.  Instead of
having a model of the pipeline in gpr_hazard.vhdl, which has to mirror
the behaviour of the real pipeline exactly, we now assign a 3-bit tag
to each value to be written back.  Subsequent instructions that need
to use the value get the tag number and stall until the value with
that tag is being written back to the register file.  (The reason for
using a tag rather than just the GPR register number is that the same
register might be written several times in successive cycles, and we
need to use the correct one of the values being written.)

For now, the forwarding paths are disabled.  That gives about a 20%
reduction in coremark performance.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>