core: Restore bypass path within execute1
This is done by having execute1 send the tag that it will write next
cycle back to control (via decode2) so that it can be matched against
the tags for the operands of the instruction being dispatched.  If
there is a match, the bypass signal for that input is set next cycle
so that execute1 will use its current output for that input.

Timing is pretty bad on the bypass path for the A7-100 at 100MHz.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>