core: Implement CTRL register
The CTRL register has a single bit called RUN.  It can be written at
bit 0 of SPR 152 and read at bits 0 and 15 of SPR 136.  It can affect
counting of events in the PMU, and on the Arty A7 it is plumbed out to
a LED.

Note that the CTRL register is accessed via two separate SPR numbers:
a privileged SPR number, which is write-only, and a non-privileged
number, which is readonly.  Hence we now need the ability to make SPRs
write-only or read-only.  Writing to a read-only SPR or reading from a
write-only SPR gives the same behaviour as accessing an unimplemented
SPR (no-op if MSR[PR]=0 and program interrupt if MSR[PR]=1).

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>