execute1: Simplify multiply instruction handling
This adds an extra cycle to the latency of the result from the
Xilinx-specific multiplier in order to make timing easier.  That means
that the overflow indication comes in the same cycle as the product,
which means that the logic in execute1 can be simplified, getting rid
of mul_finish and hold_wr_data.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>