dcache: Restructure to simplify and improve timing
This was motivated by the observation that the logic for r1.full is
complex and often seems to be a critical path, depending as it does on
the next request, including whether it is a cache hit or not, as well
as the current request being processed in the state machine.

Instead this adds r1.complete, which is only a function of the current
request, and computes r1_full <= r1.req.valid and not r1.complete.

In the past a new request would go into r1.req as soon as the previous
request was complete, whether it could be handled immediately by the
state machine or not.  This meant that there had to be a multiplexer
to select either r1.req or a new request based on the contents of r0,
depending on r1.full.  In order to eliminate that multiplexer,
requests don't go into r1.req until they are being processed by the
state machine.  A new signal 'can_do_request' indicates when the state
machine can handle the request currently in r0.

This means that whether a new request can be accepted into r0 depends
on can_do_request, and therefore so does stall_out.  Since
can_do_request depends in some cases on whether the current request is
a cache hit or not, this makes stall_out valid later than before.
However, it only goes to loadstore1, and loadstore1 effectively has a
stash buffer for its current request already, so there is no point to
implementing a stash buffer at the front of the dcache.

The address for the cache tag and data RAMs comes either from a new
request being received (d_in or m_in) or from the request in r0.
Instead of using the stall_out signal to select which, we now use a
new 'r0_repeat' signal which is true when both r0 and r1 have a valid
request; in other words, when there is a valid request in r0, we keep
reading its tags and data until r1 becomes empty.  Then if r1_full is
false but we can't process the request in r0 (can_do_request is
false), the r0_ram_ce signal goes false to preserve the output of the
RAMs as the values for the request in r0 rather than the values for a
new request.

Loads that hit on a cache line that is in the process of being
reloaded are now handled in the dcache_request process and are
considered a cache hit, rather than being handled in the state
machine.  That process also sets the forwarding controls to enable
them to get their data when necessary (i.e. when it is less than 2
cycles since the data arrived on the wishbone).

Store data from loadstore1 now goes through a separate register, not
r0.req.data, as This eliminates an unnecessary dependency of
r0.req.data on req_is_hit, and thus improves timing.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>