core: Implement vperm and vpermr instructions
And extend the vector test to test them too.

These instructions are done in two halves, left then right.  However,
the result for the left half of the vector register may need bytes
from the right half of the sources.  Therefore the first half
instruction stores its operands and waits until the second half of the
instruction appears at the input of the vector unit.  The first half
can then use the right halves of VRA and VRB as well as the left
halves which were saved away.  The right half instruction can then
also access both halves of the VRA and VRB operands.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>