core: Implement xscvdpsxds
This implements the xscvdpsxds instruction.  It is effectively a
fctidz that can access any VSX register, and sets the right-hand half
of the destination VSX register to 0.

To implement xscvdpsxds in the FPU, we make the fcti* instructions get
their 32/64-bit indication and their signed/unsigned indication from
the decode table is_32bit and is_signed fields rather than from the
instruction word.

To set the right half of the destination to zero, we add a new DRTZ
instruction-doubling type along with an OP_SETZ that executes in the
FPU and writes 0 to the destination, unless the previous operation had
the write to its destination suppressed because of a trap-enabled
invalid operation or zero-divide exception, in which case the OP_SETZ
is also suppressed.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>