vector: Implement vector extract element to GPR instructions
This implements vextu[bhw][lr]x.  These instructions are unusual in
that they read a GPR and a VR and write a GPR.  To handle this, we
implement a new repeat type, DBS, which doubles the instruction but
suppresses the RA read and the RT write on the second instruction.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>