Start removing SPRs from register file
This starts the process of removing SPRs from the register file by
moving CFAR, SRR0/1 and SPRG0-3 out of the register file and putting
them into execute1.  They are stored in a pair of small RAM arrays,
referred to as "even" and "odd".  The reason for having two arrays is
so that two values can be read and written in each cycle.  For
example, SRR0 and SRR1 can be written in parallel by an interrupt and
read in parallel by the rfid instruction.

We now have a connection from writeback to execute1 which carries the
SRR0 and partial SRR1 value for an interrupt.  Since SRR0 and SRR1 can
be written in the same cycle now, we don't need the little state
machine in writeback any more.

mfspr/mtspr to "slow" SPRs are no longer single-issue.

HSRR0/1 and HSPRG0/1 are no longer implemented.  They were never used
anyway as we have no hypervisor mode.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>