Finish off taking SPRs out of register file
With this, the register file now contains 64 entries, for 32 GPRs and
32 FPRs, rather than the 128 it had previously.  Several things get
simplified - decode1 no longer has to work out the ispr{1,2,o} values,
decode_input_reg_{a,b,c} no longer have the t = SPR case, etc.

This adds an explicit multiplexer for the mfspr result, and uses a
now-unused leg of the result multiplexer for mfspr from the RAM SPRs.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>