Restore debug access to SPRs
This provides access to the SPRs via the JTAG DMI interface.  For now
they are still accessed as if they were GPR/FPRs using the same
numbering as before (GPRs at 0 - 0x1f, SPRs at 0x20 - 0x2d, FPRs at
0x40 - 0x5f).  CFAR is now accessible this way, since it is in the SPR
RAM.  For XER, debug reads report the full value, not just the bits
that were previously stored in the register file.  The "slow" SPR mux
is not used for debug reads.

Decode2 determines on each cycle whether a debug SPR access will
happen next cycle, based on whether there is a request and whether the
current instruction (if stalled) or the next instruction (if not
stalled) accesses the SPR RAM.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>