Restore debug access to SPRs
This provides access to the SPRs via the JTAG DMI interface.  For now
they are still accessed as if they were GPR/FPRs using the same
numbering as before (GPRs at 0 - 0x1f, SPRs at 0x20 - 0x2d, FPRs at
0x40 - 0x5f).  CFAR is now accessible this way, since it is in the SPR
RAM.  For XER, the low bits are now in a register, and debug reads
report the full value, not just the bits that were previously stored
in the register file.  The "slow" SPR mux is not used for debug
reads.

The read address for the SPR RAM is somewhat timing-critical, so we
determine it in the decode2 cycle, based on whether the current
instruction (if there is a valid current instruction) needs to read
the SPR RAM and whether there is a pending debug SPR read request.
Decode2 can now update the fields of r.e relating to SPR accesses
while instruction issue is stalled by execute1, as long as the
instruction in r.e is not one that reads SPRs.

It turns out that we only need a single read address to the SPR RAM,
so this makes that simplification.

SPR decoding is now done in decode1, and map_spr() only decodes the
non-RAM SPRs.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>