Improve timing on the mfspr path
This reduces the SPR multiplexer to 8 inputs, and puts a register in
the path for the PMU SPRs since x_to_pmu.spr_val comes from a wide
multiplexer.  The no-op behaviour of privileged mfspr to an
unimplemented SPR is now done by routing c_in through logical_result
and the main result multiplexer.  The SPR RAM now has a port on the
main result multiplexer rather than going through the SPR multiplexer.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>