Store per-tag instruction addresses in execute1
Instead of carrying CIA/NIA values around in the FPU and loadstore1
pipeline, store them in a little RAM in execute1.  When an interrupt
occurs, the appropriate instruction address is fetched from the RAM
and written to SRR0.

The sc instruction is an exception to this, as it writes CIA+4 to
SRR0.  To allow for this, we have the sc instruction write next_nia to
SRR0 in the cycle when the sc is executed, and then suppress the write
to SRR0 when the interrupt is taken.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>