execute1: Restructure to separate out execution of side effects
We now have a record that represents the actions taken in executing an
instruction, and a process that computes that for the incoming
instruction.  We no longer have 'current' or 'r.cur_instr', instead
r.e is computed in the first cycle of an instruction (except for the
GPR and XER write data for slow instructions) and not reinitialized in
subsequent busy cycles.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>