execute1: Start moving instruction side-effects later
This moves several instruction side-effects one clock later, so that
in future we can defer them until the instruction completes.  The
side-effects that are moved are writes to MSR, XER, DEC, LOG_ADDR and
PMU SPRs, i-cache invalidations, and termination of execution.

In order to allow following instructions that depend on MSR or XER to
execute immediately, we keep two copies of MSR and XER, so that we can
roll back to the last committed values if an instruction gets flushed
after it is executed.

In order not to have to forward the value written to an SPR to an
immediately following mfspr to the same register, we mark mtspr to a
"slow" SPR as single-issue.  The "slow" SPRs are those in execute1
that are not in the SPR RAM, i.e. DEC and LOG_ADDR, plus the PMU SPRs.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>